intermediate commit setting up patch step invalidation
